# SourceCode/front-end/logsmart/openapi.json
{
	"openapi": "3.1.0",
	"info": { "title": "back-end", "description": "", "license": { "name": "" }, "version": "1.1.0" },
	"paths": {
...
		"/auth/company/branches": {
			"get": {
				"tags": ["Company Management"],
				"summary": "Lists all branches for the user's company.",
				"operationId": "list_branches",
				"responses": {
					"200": {
						"description": "Branches retrieved successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ListBranchesResponse" }
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			},
			"put": {
				"tags": ["Company Management"],
				"summary": "Updates an existing branch for the user's company.",
				"operationId": "update_branch",
				"requestBody": {
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/UpdateBranchRequest" } }
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Branch updated successfully",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/BranchDto" } }
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden - only company manager can update branches",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"404": {
						"description": "Branch not found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			},
			"post": {
				"tags": ["Company Management"],
				"summary": "Creates a new branch for the user's company.",
				"operationId": "create_branch",
				"requestBody": {
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/CreateBranchRequest" } }
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Branch created successfully",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/BranchDto" } }
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden - only company manager can create branches",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"BranchDto": {
				"type": "object",
				"required": ["id", "company_id", "name", "address", "created_at"],
				"properties": {
					"address": { "type": "string" },
					"company_id": { "type": "string" },
					"created_at": { "type": "string", "format": "date-time" },
					"deletion_requested_at": { "type": ["string", "null"], "format": "date-time" },
					"has_pending_deletion": { "type": "boolean" },
					"id": { "type": "string" },
					"name": { "type": "string" }
				}
			},
...
			"CreateBranchRequest": {
				"type": "object",
				"required": ["name", "address"],
				"properties": {
					"address": { "type": "string", "example": "123 Regent St, London" },
					"name": { "type": "string", "example": "London Office" }
				}
			},
...
			"ErrorResponse": {
				"type": "object",
				"required": ["error"],
				"properties": {
					"error": { "type": "string" },
					"fields": {
						"type": ["object", "null"],
						"additionalProperties": { "type": "array", "items": { "type": "string" } },
						"propertyNames": { "type": "string" }
					}
				}
			},
...
			"ListBranchesResponse": {
				"type": "object",
				"required": ["branches"],
				"properties": {
					"branches": { "type": "array", "items": { "$ref": "#/components/schemas/BranchDto" } }
				}
			},
...
			"UpdateBranchRequest": {
				"type": "object",
				"required": ["branch_id", "name", "address"],
				"properties": {
					"address": { "type": "string", "example": "123 Regent St, London" },
					"branch_id": { "type": "string", "example": "550e8400-e29b-41d4-a716-446655440000" },
					"name": { "type": "string", "example": "London Office" }
				}
			},
...
		},
		"securitySchemes": { "bearer_auth": { "type": "http", "scheme": "bearer" } }
	},
...
}